models file updated.

+++ models/demands.go
@@ -6,9 +6,9 @@ import (
 
 // Demand struct represents a demand for a specific shift
 type Demand struct {
-	ID        uint       `gorm:"primaryKey" json:"id"`
-	UserID    uint       `gorm:"column:user_id" json:"userId"`
-	ShiftID   uint       `gorm:"column:shift_id" json:"shiftId"`
+	ID        int64      `gorm:"primaryKey" json:"id"`
+	UserID    int64      `gorm:"column:user_id" json:"userId"`
+	ShiftID   int64      `gorm:"column:shift_id" json:"shiftId"`
 	StartDate time.Time  `gorm:"column:start_date" json:"startDate"`
 	EndDate   time.Time  `gorm:"column:end_date" json:"endDate"`
 	CreatedAt time.Time  `gorm:"column:created_at" json:"createdAt"`
