+++ models/user_shift_period.go
@@ -6,11 +6,11 @@ import (
 
 // UserShiftPeriod struct represents the shift period of a user
 type UserShiftPeriod struct {
-	ID        uint       `gorm:"primaryKey" json:"id"`
-	UserID    uint       `gorm:"column:user_id" json:"userId"`
+	ID        int64      `gorm:"primaryKey" json:"id"`
+	UserID    int64      `gorm:"column:user_id" json:"userId"`
 	StartDate time.Time  `gorm:"column:start_date" json:"startDate"`
 	EndDate   time.Time  `gorm:"column:end_date" json:"endDate"`
-	ShiftID   uint       `gorm:"column:shift_id" json:"shiftId"`
+	ShiftID   int64      `gorm:"column:shift_id" json:"shiftId"`
 	CreatedAt time.Time  `gorm:"column:created_at" json:"createdAt"`
 	UpdatedAt time.Time  `gorm:"column:updated_at" json:"updatedAt"`
 	DeletedAt *time.Time `gorm:"column:deleted_at" json:"-"`
